refactor: separate game dto models from mapping services
