+++ src/main/java/com/codenames/codenames/backend/game/application/GameService.java
@@ -3,8 +3,8 @@ package com.codenames.codenames.backend.game.application;
 import com.codenames.codenames.backend.clue.Clue;
 import com.codenames.codenames.backend.game.domain.GameManager;
 import com.codenames.codenames.backend.game.domain.GameManagerFactory;
-import com.codenames.codenames.backend.serialization.DataTransferObjectService;
-import com.codenames.codenames.backend.serialization.GameStateDataTransferObject;
+import com.codenames.codenames.backend.game.mapping.DataTransferObjectService;
+import com.codenames.codenames.backend.game.dto.GameStateDataTransferObject;
 import com.codenames.codenames.backend.utility.Team;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
