+++ src/main/java/com/codenames/codenames/backend/game/mapping/DataTransferObjectService.java
@@ -1,6 +1,8 @@
-package com.codenames.codenames.backend.serialization;
+package com.codenames.codenames.backend.game.mapping;
 
 import com.codenames.codenames.backend.game.dto.ClueDto;
+import com.codenames.codenames.backend.game.dto.CardDataTransferObject;
+import com.codenames.codenames.backend.game.dto.GameStateDataTransferObject;
 import com.codenames.codenames.backend.game.domain.Card;
 import com.codenames.codenames.backend.game.domain.GameManager;
 import com.codenames.codenames.backend.utility.Color;
