+++ src/main/java/com/codenames/codenames/backend/recovery/SystemStateRecoveryService.java
@@ -7,7 +7,7 @@ import com.codenames.codenames.backend.lobby.Lobby;
 import com.codenames.codenames.backend.lobby.dto.PlayerDto;
 import com.codenames.codenames.backend.lobby.services.LobbyService;
 import com.codenames.codenames.backend.recovery.snapshot.SystemSnapshot;
-import com.codenames.codenames.backend.serialization.GameStateDataTransferObject;
+import com.codenames.codenames.backend.game.dto.GameStateDataTransferObject;
 import java.util.Comparator;
 import java.util.List;
 import java.util.Map;
