+++ src/main/java/com/codenames/codenames/backend/recovery/snapshot/SystemSnapshot.java
@@ -1,7 +1,7 @@
 package com.codenames.codenames.backend.recovery.snapshot;
 
 import com.codenames.codenames.backend.lobby.dto.PlayerDto;
-import com.codenames.codenames.backend.serialization.GameStateDataTransferObject;
+import com.codenames.codenames.backend.game.dto.GameStateDataTransferObject;
 import java.util.List;
 import java.util.Map;
