+++ src/test/java/com/codenames/codenames/backend/game/application/GameServiceTest.java
@@ -13,9 +13,9 @@ import com.codenames.codenames.backend.clue.Clue;
 import com.codenames.codenames.backend.game.domain.GameManager;
 import com.codenames.codenames.backend.game.domain.GameManagerFactory;
 import com.codenames.codenames.backend.game.dto.ClueDto;
-import com.codenames.codenames.backend.serialization.CardDataTransferObject;
-import com.codenames.codenames.backend.serialization.DataTransferObjectService;
-import com.codenames.codenames.backend.serialization.GameStateDataTransferObject;
+import com.codenames.codenames.backend.game.dto.CardDataTransferObject;
+import com.codenames.codenames.backend.game.mapping.DataTransferObjectService;
+import com.codenames.codenames.backend.game.dto.GameStateDataTransferObject;
 import com.codenames.codenames.backend.utility.Color;
 import com.codenames.codenames.backend.utility.Role;
 import com.codenames.codenames.backend.utility.Team;
